Correct ctrl_port type

+++ src/pandora/db/generic_procedures/ccx_cx_commute.sql
@@ -36,7 +36,7 @@ declare
 
     start_time timestamp;
 
-    ctrl_port smallint;
+    ctrl_port int;
 
     cx_type smallint;
     cxpow_type smallint;
